rename: 폴더 구조 변경하면서 누락된 것 추가 변경
