+++ auth/src/main/java/com/devcamp/flametalk/global/util/UuidGenerator.java
@@ -1,4 +1,4 @@
-package com.devcamp.flametalk.util;
+package com.devcamp.flametalk.global.util;
 
 import java.io.Serializable;
 import java.util.Random;
